chore: refactor reset database
