+++ internal/db/remote/changes/changes.go
@@ -173,7 +173,7 @@ EOSQL
 		}
 
 		p.Send(utils.StatusMsg("Resetting database..."))
-		if err := differ.ResetDatabase(ctx, dbId, utils.ShadowDbName); err != nil {
+		if err := commit.ResetDatabase(ctx, dbId, utils.ShadowDbName); err != nil {
 			return err
 		}
