+++ internal/db/remote/commit/templates/reset.sh
